treefmt: add python linting because `ruff format` doesn't do everything
Charles point out on matrix that `ruff format` doesn't order imports,
and doesn't run lints.

This commit add a new treefmt formatter to lint the code (for now only
order imports), but more lints could be enable. I tried to enable `ALL`,
but it's too much and enforce docstring and stuff we may don't want
especially for a testsuite.